chore(react-demo): extract product.cont
